refactor: extract proxy middleware into mirror.go
